Promote volume drivers from experimental to master.
Remove volume stubs and use the experimental path as the only path.
