+++ docs/extend/index.md
@@ -1,0 +1,22 @@
+<!--[metadata]>
++++
+title = "Extend Docker"
+description = "How to extend Docker with plugins"
+keywords = ["extend, plugins, docker, documentation, developer"]
+[menu.main]
+identifier = "mn_extend"
+name = "Extend Docker"
+weight = 6
++++
+<![end-metadata]-->
+
+
+## Extending Docker
+
+Currently, you can extend Docker by adding a plugin. This section contains the following topics:
+
+* [Understand Docker plugins](plugins.md)
+* [Write a volume plugin](plugins_volumes.md)
+* [Docker plugin API](plugin_api.md)
+
+
