+++ integration-cli/docker_cli_start_volume_driver_unix_test.go
@@ -1,4 +1,3 @@
-// +build experimental
 // +build !windows
 
 package main
